docs: Consistency around admonitions
Throughout the documentation we had a bit of inconsistency around
admonitions.

As such we added few rules to keep the consistency which will improve
the UX:

* we should use `???+` instead of `!!!` so we can have a collapse block
(default open)
* for admonitions with default title, we should write the description
on a new line
* for admonitions which has a custom title as well as a description, we
should adhere to the format `<admonitions type>: <custom title>`.

E.g - for a tip admonitions with a custom title "Did you know?" and
description of "Powertools is a brilliant library"

```
???+ tip "Tip: Did you know?"
     Powertools is a brilliant library
```