feat: Add __str__ to Data Classes base DictWrapper
Add in a __str__ method to the base class used by all Data Classes.
This provides a dictionary string primarily for debugging purposes.

To construct the dictionary a helper method _str_helper recursively
works through `properties` in each instance of a Data Class.

Some Data Classes have sensitive properties, such as credentials,
which can be excluded by setting the `_sensitive_properties` attribute
(list of strings).
These properties will have their value replaced with "[SENSITIVE]",
which allows a developer to see a value arrived, but was hidden
as it is sensitive.